+++ src/Config/Config.hs
@@ -9,6 +9,7 @@ import GHC.Generics (Generic)
 import Data.Text (Text)
 import Data.Maybe (fromMaybe)
 import Data.Aeson (genericParseJSON, defaultOptions)
+import Config.ModuleTransform (transformModulePath, isValidModuleName)
 
 -- Data types for the config schema
 data Config = Config
@@ -32,6 +33,7 @@ data Template = Template
   { template :: FilePath
   , filename :: Text
   , output_by :: Text
+  , module_transform :: Maybe Text  -- New field for module transformation
   } deriving (Show, Generic, Eq)
 
 data Naming = Naming
@@ -52,7 +54,11 @@ instance FromJSON Layer where
   parseJSON = genericParseJSON defaultOptions
 
 instance FromJSON Template where
-  parseJSON = genericParseJSON defaultOptions
+  parseJSON = withObject "Template" $ \v -> Template
+    <$> v .: "template"
+    <*> v .: "filename"
+    <*> v .: "output_by"
+    <*> v .:? "module_transform"
 
 instance FromJSON Naming where
   parseJSON = genericParseJSON defaultOptions
@@ -66,4 +72,10 @@ loadConfig path = do
   result <- decodeFileEither path
   return $ case result of
     Left err -> Left $ "Failed to parse config: " ++ show err
-    Right config -> Right config 
+    Right config -> Right config
+
+-- Function to transform a module name in a template
+transformTemplateModule :: Template -> Text -> Text
+transformTemplateModule template moduleName = case module_transform template of
+  Just func -> transformModulePath func moduleName
+  Nothing -> moduleName